requirements.txt
asgiref==3.4.1
autopep8==1.6.0
black==21.12b0
click==8.0.3
Django==4.0
django-cors-headers==3.10.1
django-nested-admin==3.4.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
python-environ==0.4.54
python-monkey-business==1.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
asgiref==3.4.1
autopep8==1.6.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Django==4.0
django-cors-headers==3.10.1
django-nested-admin==3.4.0
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
python-environ==0.4.54
python-monkey-business==1.0.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
